@@ -1,5 +1,5 @@
 import React, { createContext, useState, useEffect, useContext } from 'react';
-import axios from 'axios'; 
+import axios from 'axios';
 
 const AuthContext = React.createContext();
 
@@ -10,7 +10,7 @@ export function useAuth() {
 export function AuthProvider(props) {
     const [authUser, setAuthUser] = useState(null);
     const [isLoggedIn, setIsLoggedIn] = useState(false);
-
+    
     const value = { authUser, setAuthUser, isLoggedIn, setIsLoggedIn };
 
     return (
