@@ -1,4 +1,4 @@
-import thunk from "redux-thunk";
+import { thunk } from "redux-thunk";
 import rootReducer from "../reducers";
 import { configureStore } from "@reduxjs/toolkit";
 
